requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.4
django-bootstrap4==22.3
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
easy-thumbnails==2.8.3
Pillow==9.3.0
pytz==2022.6
soupsieve==2.3.2.post1
sqlparse==0.4.3
